backend/utils/file_ops.py
#Functions for file operations like copying, moving, renaming files, and managing directories.
import os
import shutil

def copy_file(src, dest):
    ''' Copy a file from src to dest '''
    try:
        shutil.copy(src, dest)
        print(f"File copied from {src} to {dest}")
    except Exception as e:
        print(f"Error copying file from {src} to {dest}: {e}")

def move_file(src, dest):
    ''' Move a file from src to dest'''
    try:
        shutil.move(src, dest)
        print(f"File moved from {src} to {dest}")
    except Exception as e:
        print(f"Error moving file from {src} to {dest}: {e}")

def organize_files(directory, metadata):
    ''' Organize files in a directory based on metadata '''
    # Example logic: Create folders by date and move files
    for item in metadata:
        # here, you would impletment the logic to organize files
        pass